append_catalog_fields

adds helper fields to accounting.csv.

Usage

append_catalog_fields(
  path_to_ras_dbase = NULL,
  out_name = NULL,
  overwrite = FALSE,
  is_verbose = TRUE,
  HUC8_override = NULL
)

Arguments

path_to_ras_dbase

the path to the folder in which you are building your catalog, Default: NULL

out_name

the name of the csv you want to generate, Default: NULL

overwrite

flag to dictate whether or not to overwrite the out_name, should it exist. set to TRUE to delete and (re)generate, FALSE to safely exit, Default: FALSE

is_verbose

flag to determine whether print statements are suppressed, TRUE to show messages and FALSE to suppress them, Default: TRUE

HUC8_override

a path to the spatial key if you need to run this over a temp dir for eg ras2fim, Default: NULL

Value

a new csv with helper columns

Details

TRUE

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE
 RRASSLER::append_catalog_fields(path_to_ras_dbase = "G:/data/ras_catalog",out_name = "OWP_ras_model_catalog.csv",overwrite = FALSE,is_verbose = TRUE)
 }
} # }