cloud_ingest_record

Depreciated - add a file as a record in the RRASSLED structure - Depreciated

Usage

cloud_ingest_record(
  in_file = NULL,
  ras_dbase = NULL,
  root_bucket = NULL,
  code_to_place_in_source = NULL,
  proj_override = NULL,
  apply_vdat_trans = FALSE,
  is_quiet = FALSE,
  is_verbose = FALSE,
  overwrite = FALSE
)

Arguments

in_file

the path to the file on disk that we want to ingest, Default: NULL

ras_dbase

A path to an S3 bucket to write your RRASSLED catalog to., Default: NULL

root_bucket

first path in s3 url, will auto-populate as needed, Default: NULL

code_to_place_in_source

a code to place in the metadata as the owner of the model, Default: NULL

proj_override

a string to override projection information should none be found, Default: NULL

apply_vdat_trans

Should VDATUM be applied to the HEC-RAS model geometry. See https://vdatum.noaa.gov/, Default: FALSE

is_quiet

flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE

is_verbose

flag to determine whether internal print statements (i.e. cross section parsing, vdat trans, file info) are suppressed, TRUE to show these messages and FALSE to suppress them, Default: FALSE

overwrite

overwrite files if we find identical models, Default: FALSE

Value

a set of files in a newly RRASSLE'd record.

Details

the cloud version of the ingest process, depreciated in favor of local pre-RRASSL

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 }
} # }