ingest_FEMA6_BLE

helper to ingest FEMA region 6 BLE data

Usage

ingest_FEMA6_BLE(
  path_to_ras_dbase,
  HUCID,
  proj_override = NULL,
  apply_vdat_trans = FALSE,
  is_quiet = FALSE,
  is_verbose = TRUE,
  overwrite = FALSE,
  parallel_proc = TRUE,
  free_treads = 2,
  clean = FALSE,
  opt_local_path = NULL
)

Arguments

path_to_ras_dbase

The path to the folder in which you are building your catalog, Default: NULL

HUCID

string to huc8

proj_override

a CRS string to apply should a projection not be found, Default: NULL

apply_vdat_trans

a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE

is_quiet

flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE

is_verbose

flag to determine whether internal print statements (i.e. cross section parsing, vdat trans, file info) are suppressed, TRUE to show these messages and FALSE to suppress them, Default: TRUE

overwrite

overwrite files if we find identical models, Default: FALSE

parallel_proc

Flag to determine if this should this parallel process, will check for enough free cores and boot this back if it exceeds available resources. Will suppress all intermediate messages if active, Default: TRUE

free_treads

number of threads to leave free if parallel processing, Default: 2

clean

number of threads to leave free if parallel processing, Default: 2

opt_local_path

PARAM_DESCRIPTION, Default: NULL

Value

a RRASSLE'd catalog of models or added desired HUC8 models

Details

As one of the best centralized and accessible databases, the FEMA region 6 BLE data are perfect candidates to RRASSL. This provides a wrapper around BLE scrapers and the ingest_into_database function.

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 RRASSLER::ingest_FEMA6_BLE(path_to_ras_dbase = "G:/data/ras_catalog/","12090301",proj_override = "EPSG:2277",apply_vdat_trans = FALSE,is_quiet = FALSE,is_verbose = FALSE,overwrite = FALSE,parallel_proc = TRUE,free_treads = 2)
 }
} # }