make_xs_hyfab_comp

testing function to crosswalk cross sections to flowline distance

Usage

make_xs_hyfab_comp(
  path_to_ras_dbase = NULL,
  line_select_subset = NULL,
  hf_lines,
  overwrite = FALSE,
  is_verbose = TRUE
)

Arguments

path_to_ras_dbase

A path to a RRASSLED directory to, Default: NULL

line_select_subset

A subset of the lines you want to crosswalk, Default: NULL

hf_lines

the hydrofabric network you want to crosswalk to

overwrite

overwrite outputs, Default: FALSE

is_verbose

flag to determine whether internal print statements (i.e. cross section parsing, vdat trans, file info) are suppressed, TRUE to show these messages and FALSE to suppress them, Default: TRUE

Value

testing

Details

DETAILS

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 make_xs_hyfab_comp(path_to_ras_dbase = NULL,line_select_subset = NULL,hf_lines,overwrite = FALSE,is_verbose = TRUE)
 }
} # }