merge_rivers

Merges river lines

Usage

merge_rivers(path_to_ras_dbase, is_verbose = TRUE)

Arguments

path_to_ras_dbase

The path to the folder in which you are building your catalog, Default: NULL

is_verbose

flag to determine whether print statements are shown - TRUE to show messages - FALSE to skip non-critical ones, Default: TRUE

Value

Merged HEC-RAS flowpath conceptual equivalents

Details

DETAILS

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 }
} # }