process_ras_g_to_xyz

process a ras g file into xyz format

Usage

process_ras_g_to_xyz(
  geom_path,
  units,
  proj_string,
  in_epoch_override = as.integer(as.POSIXct(Sys.time())),
  out_epoch_override = as.integer(as.POSIXct(Sys.time())),
  vdat = FALSE,
  quiet = FALSE
)

Arguments

geom_path

path to a file to parse

units

units found in the project, "English Units" or "SI Units"

proj_string

a projection string to apply

in_epoch_override

vdatum parameter input epoch, Default: as.integer(as.POSIXct(Sys.time()))

out_epoch_override

vdatum parameter output epoch, Default: as.integer(as.POSIXct(Sys.time()))

vdat

a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE, Default: FALSE

quiet

flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE

Value

a point database and notes about processing

Details

DETAILS

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 #EXAMPLE1
 # g_path <- "./inst/extdata/sample_ras/FEMA-R6-BLE-sample-dataset/12090301/12090301_models/Model/Alum Creek-Colorado River/ALUM 006/ALUM 006.g01"
 pts <- process_ras_g_to_xyz(geom_path = g_path,units = "English Units",proj_string = "EPSG:2277",vdat = FALSE,quiet = FALSE)
 }
} # }