refresh_master_files

remerge individual files into spatial model key

Usage

refresh_master_files(path_to_ras_dbase, is_verbose = TRUE, overwrite = TRUE)

Arguments

path_to_ras_dbase

The path to the folder in which you are building your catalog, is also location agnostic (disk or cloud), Default: NULL

is_verbose

flag to determine whether print statements are shown - TRUE to show messages - FALSE to skip non-critical ones, Default: TRUE

overwrite

flag to determine whether the catalog is either overwritten (overwrite = TRUE), or a duplicated catalog is copied and appended with the date (overwrite = FALSE) as requested [here](https://github.com/NOAA-OWP/RRASSLER/issues/6), Default: TRUE

Value

updated master index files including accounting.csv, point_database.parque, XS.fgb, and model_footprints.fgb

Details

TRUE

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
 # ras_dbase <- file.path("~/data/ras_catalog/")
 ras_dbase <- file.path("./inst/extdata/sample_output/ras_catalog/")

 RRASSLER::refresh_master_files(path_to_ras_dbase = ras_dbase,is_verbose = TRUE)

 RRASSLER::refresh_master_files(path_to_ras_dbase = "s3://ras-models/",is_verbose = TRUE, overwrite = FALSE)
 }
} # }