ingest_FEMA6_BLE

ingest or append a new set of HEC-RAS models from FEMA region 6 Base Level Engineering (BLE) into your database

ingest or append a new set of HEC-RAS models into your database

Usage

ingest_FEMA6_BLE(
  path_to_ras_dbase,
  HUC8,
  proj_overwrite = NULL,
  vdat_trans = FALSE,
  quiet = TRUE,
  chatty = FALSE,
  ping_me = NULL,
  quick_check = FALSE,
  quick_hull = FALSE,
  overwrite = FALSE,
  refresh = TRUE
)

ingest_into_database(
  path_to_ras_dbase,
  top_of_dir_to_scrape,
  code_to_place_in_source,
  proj_overwrite = NULL,
  vdat_trans = FALSE,
  quiet = TRUE,
  chatty = FALSE,
  ping_me = NULL,
  quick_check = FALSE,
  quick_hull = FALSE,
  overwrite = FALSE,
  refresh = TRUE
)

Arguments

path_to_ras_dbase

The path to the folder in which you are building your catalog, Default: NULL

HUC8

The HUC8 string you'd like to ingest

proj_overwrite

an EPSG string to apply should a projection not be found, Default: NULL

vdat_trans

a flag to dictate whether or not to apply a vdatum transformation, TRUE to apply, FALSE to skip, Default: FALSE

quiet

flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE

chatty

flag to dictate whether print statements from within the extraction are suppressed, TRUE to show messages and FALSE to suppress them, Default: FALSE

ping_me

a string with an email used to send emails after processing. Uses gmailr and requires config, Default: NULL

quick_check

a flag to dictate whether or not to perform a quick check to see if a model has already been processed based on the raw name of the file. Useful in reingesting the same directory after an error but not recommended otherwise. TRUE to see if a name matches and skips, FALSE to process all the way though, Default: FALSE

quick_hull

a flag to dictate how tightly you want the footprints wrapped, TRUE uses just end points of the linestrings to place the model, FALSE uses the full point database, Default: FALSE

overwrite

not currently implemented, Default: FALSE

refresh

flag to dictate whether or not to recollate spatial database after ingest process. FALSE to skip, TRUE to regenerate, Default: TRUE

top_of_dir_to_scrape

The path to the top of the directory which you want to ingest

code_to_place_in_source

a string which encodes the owner or maintainer of that model

Value

a RRASSLED catalog of HEC-RAS models

a RRASSLED catalog of HEC-RAS models

Details

DETAILS

DETAILS

See also

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}
if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}