map_library

a function to help apply geographic context to the RRASSLED catalog

Usage

map_library(
  path_to_ras_dbase,
  AOI_to_map = NULL,
  name = "model_map",
  plot_lines = FALSE,
  chart_lines = FALSE,
  refresh = FALSE,
  quiet = TRUE
)

Arguments

path_to_ras_dbase

The path to the folder in which you are building your catalog, Default: NULL

AOI_to_map

An AOI pass though to subset the catalog down into a smaller area to add the context too, Default: NULL

name

A name for the map files to create, Default: 'model_map'

plot_lines

draw lines on the map as opposed to just the footprints, Default: FALSE

chart_lines

Add cross section click graphs to the plot lines. Will break if area is too large, Default: FALSE

refresh

flag to dictate whether or not to recollate spatial database prior to mapping. FALSE to skip, TRUE to regenerate, Default: TRUE

quiet

flag to determine whether print statements are suppressed, TRUE to suppress messages and FALSE to show them, Default: FALSE

Value

OUTPUT_DESCRIPTION

Details

DETAILS

See also

Examples

if (FALSE) {
if(interactive()){
 #EXAMPLE1
 }
}